Extract mdoc ECDSA/sign1 helpers from keystore into utils
